+++ src/utils/logger.ts
@@ -1,0 +1,5 @@
+import { Logger } from "@chrismessina/raycast-logger";
+
+export const logger = new Logger({
+  prefix: "[fly]",
+});